+++ vector.h
@@ -16,8 +16,8 @@ Vector* vector_create(size_t);
 void vector_delete(Vector*);
 
 void* vector_get(Vector*, size_t);
-int vector_in(Vector*, void*);
-int vector_length(Vector*);
+size_t vector_in(Vector*, void*);
+size_t vector_length(Vector*);
 void* vector_pop(Vector*);
 void vector_push(Vector*, void*);
 void vector_reverse(Vector*);
